feat: Tensorflow Seldon servers integration tests (#152)
https://github.com/canonical/seldon-core-operator/issues/151

Summary of changes:
- Added integration tests for testing Tensorflow Seldon servers.
- Moved test data files into `tests/data/`.
- Modified code to read from files when necessary to avoid having large   test data structures in the code.
- Reduced resource requirements to test execution of tests in GH   runners.
- Moved test assets for better organization.
- Added clean up of applications and relation after every test to   free up resources.
- Split seldon server tests into separate file.
- Re-organized integration tests to conserve resources.
- Added longer timeouts to tenacity functions.
- Re-factored test utility functions.
- Added requests to ephemeral-storage to tests.
- Reorgannized test to be aligned with common ci
