+++ tests/assets/crs/serve-simple-v1.yaml
