+++ tests/assets/crs/serve-simple-v1alpha2.yml
