+++ tests/assets/crs/tensorflow-serving.yaml
@@ -1,0 +1,35 @@
+# https://docs.seldon.io/projects/seldon-core/en/latest/servers/tensorflow.html
+apiVersion: machinelearning.seldon.io/v1
+kind: SeldonDeployment
+metadata:
+  name: tfserving
+spec:
+  name: mnist
+  resources:
+    requests:
+      memory: "64Mi"
+      cpu: "125m"
+    limits:
+      memory: "128Mi"
+      cpu: "125m"
+  predictors:
+  - graph:
+      children: []
+      implementation: TENSORFLOW_SERVER
+      modelUri: gs://seldon-models/tfserving/mnist-model
+      name: classifier
+      parameters:
+        - name: signature_name
+          type: STRING
+          value: predict_images
+        - name: model_name
+          type: STRING
+          value: classifier
+        - name: model_input
+          type: STRING
+          value: images
+        - name: model_output
+          type: STRING
+          value: scores
+    name: default
+    replicas: 1
