+++ tests/assets/crs/xgboost-v2.yaml
