+++ tox.ini
@@ -74,7 +74,9 @@ deps =
 description = Run unit tests
 
 [testenv:integration]
-commands = pytest -v --tb native --asyncio-mode=auto {[vars]tst_path}integration --log-cli-level=INFO -s {posargs}
+commands =
+    pytest -vv --tb native --asyncio-mode=auto {[vars]tst_path}integration --log-cli-level=INFO -s {posargs}
 deps =
     -r requirements-integration.txt
 description = Run integration tests
+
